# docs/api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GOAT AI",
    "description": "Simon Business School Strategic Intelligence API",
    "version": "1.2.0"
  },
  "paths": {
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List available Ollama models",
        "description": "Return the list of locally available Ollama model names.",
        "operationId": "list_models_route_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/capabilities": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Read capabilities for one Ollama model",
        "description": "Return Ollama-reported capabilities for the selected model.",
        "operationId": "get_model_capabilities_route_api_models_capabilities_get",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCapabilitiesResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List persisted chat sessions",
        "description": "Return session metadata rows for the history sidebar.",
        "operationId": "list_history_api_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistorySessionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "history"
        ],
        "summary": "Delete all persisted chat sessions",
        "description": "Delete all persisted sessions and related conversation rows.",
        "operationId": "delete_all_history_api_history_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{session_id}": {
      "patch": {
        "tags": [
          "history"
        ],
        "summary": "Rename one persisted chat session",
        "description": "Rename one persisted session.",
        "operationId": "rename_history_session_route_api_history__session_id__patch",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistorySessionRenameRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get one persisted chat session",
        "description": "Return one full persisted session including all messages.",
        "operationId": "get_history_session_api_history__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistorySessionDetailResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "history"
        ],
        "summary": "Delete one persisted chat session",
        "description": "Delete one persisted session.",
        "operationId": "delete_history_session_api_history__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Detail",
            "description": "Human-readable message or structured validation errors."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Echo of X-Request-ID when the server generated or accepted one."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "Standard JSON error envelope (Phase 13 §13.0)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistorySessionDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Schema Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": ""
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/HistorySessionMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "chart_spec": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Spec"
          },
          "file_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistorySessionFileContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "knowledge_documents": {
            "items": {
              "$ref": "#/components/schemas/HistorySessionKnowledgeDocument"
            },
            "type": "array",
            "title": "Knowledge Documents"
          },
          "workspace_outputs": {
            "items": {
              "$ref": "#/components/schemas/WorkbenchWorkspaceOutputPayload"
            },
            "type": "array",
            "title": "Workspace Outputs",
            "description": "Durable workbench outputs linked to this session."
          },
          "chart_data_source": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(uploaded|demo|none)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Data Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "model",
          "schema_version",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "HistorySessionDetailResponse",
        "description": "Body for GET /api/history/{session_id}."
      },
      "HistorySessionFileContext": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "HistorySessionFileContext",
        "description": "Normalized stored file-context metadata for one session."
      },
      "HistorySessionKnowledgeDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "mime_type"
        ],
        "title": "HistorySessionKnowledgeDocument",
        "description": "Indexed knowledge document attached to a chat session."
      },
      "HistorySessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/HistorySessionSummary"
            },
            "type": "array",
            "title": "Sessions"
          }
        },
        "type": "object",
        "required": [
          "sessions"
        ],
        "title": "HistorySessionListResponse",
        "description": "Body for GET /api/history."
      },
      "HistorySessionMessage": {
        "properties": {
          "role": {
            "type": "string",
            "pattern": "^(user|assistant|system)$",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "image_attachment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Attachment Ids",
            "description": "Vision image attachment ids stored for this user turn."
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ChatArtifact"
            },
            "type": "array",
            "title": "Artifacts",
            "description": "Downloadable generated files associated with this assistant turn."
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "HistorySessionMessage",
        "description": "Stored message row for persisted session history."
      },
      "HistorySessionRenameRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "HistorySessionRenameRequest",
        "description": "Body for PATCH /api/history/{session_id}."
      },
      "HistorySessionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Schema Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "model",
          "schema_version",
          "created_at",
          "updated_at"
        ],
        "title": "HistorySessionSummary",
        "description": "Lightweight session metadata for sidebar/history list views."
      },
...
      "ModelCapabilitiesResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "supports_tool_calling": {
            "type": "boolean",
            "title": "Supports Tool Calling"
          },
          "supports_chart_tools": {
            "type": "boolean",
            "title": "Supports Chart Tools"
          },
          "supports_vision": {
            "type": "boolean",
            "title": "Supports Vision"
          },
          "supports_thinking": {
            "type": "boolean",
            "title": "Supports Thinking"
          },
          "context_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Length",
            "description": "Ollama-reported context window in tokens, when discoverable from /api/show."
          }
        },
        "type": "object",
        "required": [
          "model",
          "capabilities",
          "supports_tool_calling",
          "supports_chart_tools",
          "supports_vision",
          "supports_thinking"
        ],
        "title": "ModelCapabilitiesResponse",
        "description": "Capabilities for one Ollama model."
      },
      "ModelsResponse": {
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelsResponse",
        "description": "Body for GET /api/models."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}